+++ e_metrobus/navigation/views.py
@@ -1,6 +1,6 @@
 import posthog
 
-from django.shortcuts import redirect, Http404, get_object_or_404
+from django.shortcuts import redirect, Http404, get_object_or_404, HttpResponse
 from django.views.generic import TemplateView
 
 from e_metrobus.navigation import chart
@@ -88,9 +88,6 @@ class DashboardView(CheckStationsMixin, NavigationView):
     def get(self, request, *args, **kwargs):
         if questions.all_questions_answered(request.session):
             return redirect("navigation:finished_quiz")
-        if "first_time" not in request.session:
-            request.session["first_time"] = False
-            kwargs["first_time"] = True
         return super(DashboardView, self).get(request, *args, **kwargs)
 
     def get_context_data(self, **kwargs):
@@ -182,6 +179,9 @@ class ComparisonView(CheckStationsMixin, NavigationView):
             [int(route_data[vehicle].co2) for vehicle in chart_order]
         )
         context["info_table"] = widgets.InfoTable()
+        if "first_time" not in self.request.session:
+            self.request.session["first_time"] = False
+            context["first_time"] = True
         return context
 
 
@@ -373,7 +373,18 @@ class LegalView(NavigationView):
     def get_context_data(self, **kwargs):
         context = super(LegalView, self).get_context_data(**kwargs)
         context["info_table"] = widgets.InfoTable()
+        context["bug"] = kwargs.get(
+            "bug", forms.BugForm(initial={"type": models.Bug.TECHNICAL}),
+        )
         return context
+
+    def post(self, request, **kwargs):
+        bug = forms.BugForm(request.POST)
+        if bug.is_valid():
+            bug.save()
+        else:
+            return self.render_to_response(self.get_context_data(bug=bug))
+        return redirect("navigation:legal")
 
 
 class QuestionsAsTextView(NavigationView):
@@ -399,6 +410,8 @@ class LandingPageView(PosthogMixin, TemplateView):
         context = super(LandingPageView, self).get_context_data(**kwargs)
         if "visited" in self.request.GET:
             context["visited"] = True
+        if "privacy" in self.request.session:
+            context["privacy_accepted"] = True
         return context
 
 
@@ -433,26 +446,14 @@ class FeedbackView(PosthogMixin, TemplateView):
         return redirect("navigation:finished_quiz")
 
 
-class BugView(NavigationView):
-    template_name = "navigation/bug.html"
-    footer_links = {
-        "info": {"enabled": True},
-        "dashboard": {"enabled": True},
-        "leaf": {"enabled": True},
-        "results": {"enabled": True},
-    }
+class TourView(NavigationView):
+    template_name = "navigation/tour.html"
 
     def get_context_data(self, **kwargs):
-        context = super(BugView, self).get_context_data(**kwargs)
-        context["bug"] = kwargs.get(
-            "bug", forms.BugForm(initial={"type": models.Bug.TECHNICAL}),
-        )
+        context = super(TourView, self).get_context_data(**kwargs)
         return context
 
-    def post(self, request, **kwargs):
-        bug = forms.BugForm(request.POST)
-        if bug.is_valid():
-            bug.save()
-        else:
-            return self.render_to_response(self.get_context_data(bug=bug))
-        return redirect("navigation:dashboard")
+
+def accept_privacy_policy(request):
+    request.session["privacy"] = True
+    return HttpResponse()